requirements/local.txt
-r base.txt
coverage==3.7.1
django-debug-toolbar==1.2
model_mommy==1.2.0